✨ 유저 게시글 헤더 컴포넌트의 유저정보와 버튼 컴포넌트 분리
