+++ components/Posts/UserPost/UserPostHeader/PostHeaderButtons.tsx/index.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import { PostType } from 'types/post';
+
+const PostHeaderButtons = ({ post }: { post: PostType }) => {
+  return <ButtonWrapper></ButtonWrapper>;
+};
+
+export default PostHeaderButtons;
+
+const ButtonWrapper = styled.div`
+  width: 300px;
+  background-color: coral;
+`;
